delete indexjs and create appjs for prerender
